+++ src/main/kotlin/com/github/warningimhack3r/intellijshadcnplugin/listeners/ToolWindowListener.kt
@@ -1,28 +1,28 @@
 package com.github.warningimhack3r.intellijshadcnplugin.listeners
 
-import com.github.warningimhack3r.intellijshadcnplugin.ui.ISPWindowContents
+import com.github.warningimhack3r.intellijshadcnplugin.ui.ISPPanelPopulator
 import com.intellij.openapi.diagnostic.logger
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.wm.ToolWindowManager
 import com.intellij.openapi.wm.ex.ToolWindowManagerListener
 
 class ToolWindowListener(private val project: Project) : ToolWindowManagerListener {
-    private val log = logger<ToolWindowListener>()
-    private val toolWindowId = "shadcn/ui"
+    companion object {
+        private val log = logger<ToolWindowListener>()
+        private const val TOOL_WINDOW_ID = "shadcn/ui"
+    }
     private var isToolWindowOpen: Boolean? = null
 
     override fun stateChanged(toolWindowManager: ToolWindowManager) {
         val previousState = isToolWindowOpen
-        isToolWindowOpen = toolWindowManager.getToolWindow(toolWindowId)?.isVisible ?: false
+        isToolWindowOpen = toolWindowManager.getToolWindow(TOOL_WINDOW_ID)?.isVisible ?: false
         if (previousState == false && isToolWindowOpen == true) {
             log.info("Tool window was closed and is now open, updating contents")
-            toolWindowManager.getToolWindow(toolWindowId)?.contentManager?.getContent(0)?.let {
+            toolWindowManager.getToolWindow(TOOL_WINDOW_ID)?.contentManager?.getContent(0)?.let {
                 with(it.component) {
                     if (components.isEmpty()) return@let
                     log.info("Removing old contents and adding new ones")
-                    remove(components[0])
-                    add(ISPWindowContents(project).panel())
-                    revalidate()
+                    ISPPanelPopulator(project).populateToolWindowPanel(this)
                 }
             }
         }